# PhoneBook.cs
namespace pb
{
    public class PhoneBook
    {
        public int ID { get; set; }
        public string PhoneNumber { get; set; } = string.Empty;
        public string Address { get; set; } = string.Empty;
        public string Name { get; set; } = string.Empty;
    }
}